pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "rsi-econ"
version = "0.1.0"
description = "Trusted bridge, canonical state, and local-only Stage 3 seed-agent substrate for RSI."
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.10,<4.0",
    "fastapi>=0.115,<1.0",
    "httpx>=0.27,<1.0",
    "pydantic>=2.7,<3.0",
    "uvicorn>=0.30,<1.0",
]

[project.optional-dependencies]
agent = [
    "pytest>=8.0,<9.0",
]
browser = [
    "playwright>=1.52,<2.0",
]
dev = [
    "pytest>=8.0,<9.0",
    "pyyaml>=6.0,<7.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
markers = [
    "fast: low-runtime invariant checks intended to run on every change",
]

[tool.setuptools.packages.find]
include = ["shared*", "trusted*", "untrusted*"]
